requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==3.2
django-seed==0.3.1
djangorestframework==3.14.0
Faker==15.0.0
psycopg2==2.9.3
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
toposort==1.7
